[Hammer] Moved more common functionality into loadtest package
This is a rough approximation of what a library version of the hammer
could look like. The remaining code in the outer `main` package is now
the tlog-tiles specific stuff. The code in the `loadtest` package should
be configurable to use with Static CT.
